requirements.txt
pandas
numpy
scikit-learn
streamlit
requests
nltk
scipy
python-dotenv